refactor: useMentoringApplicationList 훅으로 분리 #702
